Rename libta to libautomata
The library includes both code for timed automata and alternating timed
automata. Its top namespace is also called automata. Rename accordingly.
