+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(ta)
+add_subdirectory(automata)
 add_subdirectory(mtl)
 add_subdirectory(mtl_ata_translation)
 add_subdirectory(utilities)
