+++ src/automata/CMakeLists.txt
