+++ src/automata/automata.cpp
@@ -1,7 +1,7 @@
 /***************************************************************************
- *  ata_formula.cpp - Alternating Timed Automata Formulas
+ *  automata.cpp - Generic automata definitions
  *
- *  Created: Thu 28 May 2020 14:41:01 CEST 14:41
+ *  Created: Thu 28 May 2020 15:46:12 CEST 15:46
  *  Copyright  2020  Till Hofmann <hofmann@kbsg.rwth-aachen.de>
  ****************************************************************************/
 
@@ -18,23 +18,21 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#include <ta/ata_formula.h>
-#include <ta/automata.h>
-
-#include <range/v3/algorithm.hpp>
-#include <range/v3/view.hpp>
+#include <automata/automata.h>
 
 namespace automata {
-namespace ata {
 
-template class Formula<std::string>;
-template class TrueFormula<std::string>;
-template class FalseFormula<std::string>;
-template class LocationFormula<std::string>;
-template class ClockConstraintFormula<std::string>;
-template class ConjunctionFormula<std::string>;
-template class DisjunctionFormula<std::string>;
-template class ResetClockFormula<std::string>;
+bool
+is_satisfied(const ClockConstraint &constraint, const Time &valuation)
+{
+	return std::visit([&](auto &&c) { return c.is_satisfied(valuation); }, constraint);
+}
 
-} // namespace ata
+std::ostream &
+operator<<(std::ostream &os, const ClockConstraint &constraint)
+{
+	std::visit([&os](auto &&c) { os << c; }, constraint);
+	return os;
+}
+
 } // namespace automata
