+++ src/automata/include/automata/ata_formula.h
