+++ src/automata/include/automata/automata.h
