+++ src/automata/include/automata/ta.h
