+++ src/mtl_ata_translation/translator.cpp
@@ -18,12 +18,12 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
+#include "automata/ata_formula.h"
+#include "automata/automata.h"
 #include "mtl/MTLFormula.h"
-#include "ta/ata_formula.h"
-#include "ta/automata.h"
 
+#include <automata/ata.h>
 #include <mtl_ata_translation/translator.h>
-#include <ta/ata.h>
 
 #include <memory>
 #include <stdexcept>
