+++ test/test_print_ata_formula.cpp
@@ -18,9 +18,9 @@
  *  Read the full text in the LICENSE.md file.
  */
 
-#include "ta/automata.h"
+#include "automata/automata.h"
 
-#include <ta/ata_formula.h>
+#include <automata/ata_formula.h>
 
 #include <catch2/catch.hpp>
 #include <functional>
