+++ test/test_ta.cpp
@@ -18,7 +18,7 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#include <ta/ta.h>
+#include <automata/ta.h>
 
 #include <catch2/catch.hpp>
 #include <functional>
